Fix bug where file logs weren't being written.
Still getting a handle on the intricacies of the logging package...
Although the file loggers were set to not propagate messages up the stack,
they were still getting their logging level from the config of the root
logger. Need to combine setting non-propagation and explicit DEBUG level
for the desired behaviour.